Add genric H5 FileType added new example.
Add Generic H5 FileType suggested by Nat. Add example app of having a subset of the complete functionality exposed to the tool contract.

By design, the layer is leaky to avoid tryting to create a "complete" mapping of the argparse -> tool contract. Most tools in pipelines have a small minimal set of options and inputs.