DEP-117: create tmp_dir on cluster nodes if needed
This is a potential fix for creating the tmp_dir directory pointed to by the
userdata/tmp_dir symbolic link on cluster nodes where the destination
directory may not exist.  This will only handle the cases that use the tmp_dir
via the TMP_FILE and TMP_DIR resources sepcified by the tool contract.

The approach is to determine the realpath of the tmp file or dir, then try to
make the parent (dirname) directory.  The realpath should resolve to the
expected path even if the symlink is broken.  If the parent directory still
does not exist after trying to make it, an exception will be raised.  If it
exists (i.e. it already existed or just got created), the tmp_dir should exist
and the rest of the code should execute normally.

See DEP-117 for more details.